backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
openai==1.10.0
numpy==1.26.3
pydantic==2.5.3
psycopg2-binary==2.9.9
